pyproject.toml
[project]
name = "ml-ts-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lightning>=2.5.2",
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
    "mlflow>=2.8.0",
]
